+++ app/controllers/application_controller.rb
@@ -16,6 +16,7 @@ class ApplicationController < ActionController::Base # rubocop:disable Style/Doc
     %r{^/tournaments$},
     %r{^/tournaments/[0-9]+$},
     %r{^/tournaments/[0-9]+/players$},
+    %r{^/tournaments/[0-9]+/registration$},
     %r{^/tournaments/[0-9]+/rounds$},
     %r{^/tournaments/[0-9]+/rounds/[0-9]+$},
     %r{^/tournaments/[0-9]+/rounds/[0-9]+/pairings$},
