+++ app/frontend/players/Players.svelte
@@ -266,7 +266,6 @@
       <PlayerForm
         player={newPlayer}
         tournament={data.tournament}
-        tournamentPolicies={data.tournamentPolicies}
         identityNames={identityNames ?? { corp: [], runner: [] }}
         organizerView={true}
         savedCallback={newPlayerSavedCallback}
@@ -434,7 +433,6 @@
         <PlayerForm
           {player}
           tournament={data.tournament}
-          tournamentPolicies={data.tournamentPolicies}
           identityNames={identityNames ?? { corp: [], runner: [] }}
           organizerView={true}
           savedCallback={loadData}
