+++ app/views/beta/tournaments/registration.html.slim
@@ -1,0 +1,4 @@
+#registration_anchor.col-12 data-tournament-id="#{@tournament.id}" data-user-id="#{current_user&.id}"
+
+= content_for :head do
+  = vite_typescript_tag 'registration', 'data-turbolinks-track': 'reload'
